Updated MRT_ROS_Gazebo_Loop: 1) Added updateFullModalState but not functional yet. 2) Updated getCurrentObservation by depricating linkStateCallback to get base state. Instead transform btw base and world is used.

+++ ocs2_robotic_examples/ocs2_mobile_manipulator_ros/src/MobileManipulatorGazeboMRT.cpp
@@ -133,9 +133,9 @@ int main(int argc, char** argv)
     std::cout << i << ": " << manipulatorModelInfo.dofNames[i] << std::endl;
   }
 
-  int modalMode = 1;
-  int baseStateDim = 3;
-  int armStateDim = 6;
+  size_t modalMode = 1;
+  size_t baseStateDim = 3;
+  size_t armStateDim = 6;
 
   // MRT
   MRT_ROS_Interface mrt(robotName, modalMode);
